Remove unused pointcuts and move last one to the correct namespace
