+++ src/Go/Aop/Pointcut/DynamicMethodMatcherPointcut.php
@@ -6,13 +6,13 @@
  * @license       http://www.opensource.org/licenses/mit-license.php The MIT License
  */
 
-namespace Go\Aop\Support;
+namespace Go\Aop\Pointcut;
 
-use Go\Aop\MethodMatcher;
 use Go\Aop\Pointcut;
 use Go\Aop\ClassFilter;
 use Go\Aop\PointFilter;
 use Go\Aop\TrueClassFilter;
+use Go\Aop\Support\DynamicMethodMatcher;
 
 /**
  * Convenient superclass when we want to force subclasses to implement MethodMatcher interface,
